mavlink: move HIGH_LATENCY2 stream to streams folder in a single header
